Restructure and some additions to Views
