--- java/src/use_case/main_menu/SwitchToGalleryInputBoundary.java
@@ -1,5 +1,0 @@
-package use_case.main_menu;
-
-public interface SwitchToGalleryInputBoundary {
-    void execute();
-}
