--- java/src/use_case/main_menu/SwitchToGalleryOutputBoundary.java
@@ -1,7 +1,0 @@
-package use_case.main_menu;
-
-public interface SwitchToGalleryOutputBoundary {
-    void prepareSuccessView();
-
-    void prepareFailView(String error);
-}
