+++ src/use_case/main_menu/MainMenuInputBoundary.java
@@ -1,0 +1,5 @@
+package use_case.main_menu;
+
+public interface MainMenuInputBoundary {
+    void execute();
+}
