+++ src/use_case/main_menu/SwitchToGalleryInputBoundary.java
@@ -1,0 +1,5 @@
+package use_case.main_menu;
+
+public interface SwitchToGalleryInputBoundary {
+    void execute();
+}
