+++ src/use_case/main_menu/SwitchToGalleryInteractor.java
@@ -1,0 +1,12 @@
+package use_case.main_menu;
+
+public class SwitchToGalleryInteractor implements SwitchToGalleryInputBoundary{
+    final SwitchToGalleryOutputBoundary switchToGalleryPresenter;
+
+    public SwitchToGalleryInteractor(SwitchToGalleryOutputBoundary switchToGalleryPresenter) {
+        this.switchToGalleryPresenter = switchToGalleryPresenter;
+    }
+    public void execute() {
+        switchToGalleryPresenter.prepareSuccessView();
+    }
+}
